@@ -12,26 +12,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^17.0.5",
-    "@angular/compiler": "^17.0.5",
-    "@angular/core": "^17.0.5",
-    "@angular/platform-browser": "^17.0.5",
-    "@angular/platform-browser-dynamic": "^17.0.5",
-    "@angular/router": "^17.0.5",
+    "@angular/common": "^18.2.11",
+    "@angular/compiler": "^18.2.11",
+    "@angular/core": "^18.2.11",
+    "@angular/platform-browser": "^18.2.11",
+    "@angular/platform-browser-dynamic": "^18.2.11",
+    "@angular/router": "^18.2.11",
     "rxjs": "~7.8.0",
     "tslib": "^2.6.2",
     "unleash-proxy-client": "^3.2.0",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.5",
-    "@angular-eslint/builder": "17.1.1",
-    "@angular-eslint/eslint-plugin": "17.1.1",
-    "@angular-eslint/eslint-plugin-template": "17.1.1",
-    "@angular-eslint/schematics": "17.1.1",
-    "@angular-eslint/template-parser": "17.1.1",
-    "@angular/cli": "~17.0.5",
-    "@angular/compiler-cli": "^17.0.5",
+    "@angular-devkit/build-angular": "^18.2.11",
+    "@angular-eslint/builder": "^18.4.0",
+    "@angular-eslint/eslint-plugin": "^18.4.0",
+    "@angular-eslint/eslint-plugin-template": "^18.4.0",
+    "@angular-eslint/schematics": "^18.4.0",
+    "@angular-eslint/template-parser": "^18.4.0",
+    "@angular/cli": "~18.2.11",
+    "@angular/compiler-cli": "^18.2.11",
     "@types/jasmine": "~5.1.4",
     "@typescript-eslint/eslint-plugin": "^6.13.1",
     "@typescript-eslint/parser": "^6.13.1",
@@ -42,8 +42,8 @@
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "ng-packagr": "^17.0.2",
-    "typescript": "~5.2.2"
+    "ng-packagr": "^18.2.1",
+    "typescript": "~5.5.4"
   },
   "description": "Angular wrapper for [unleash-proxy-client](https://github.com/Unleash/unleash-proxy-client-js).",
   "main": "index.js",
