@@ -1,9 +1,9 @@
 {
   "name": "@karelics/angular-unleash-proxy-client",
-  "version": "17.0.0",
+  "version": "18.0.0",
   "peerDependencies": {
-    "@angular/common": ">=17.0.0",
-    "@angular/core": ">=17.0.0",
+    "@angular/common": ">=18.0.0",
+    "@angular/core": ">=18.0.0",
     "unleash-proxy-client": ">=3.0.0"
   },
   "dependencies": {
